Simplifying instruction operands:
- eliminating Operand class
- eliminating Value15 class

+++ src/main/java/lv/volkovs/myvm/instruction/impl/Jf.java
@@ -22,7 +22,7 @@ public class Jf implements Instruction {
 
     @Override
     public int execute(Memory memory, int pointer) {
-        if (memory.constOrRegister(zero).toInt() == 0) {
+        if (memory.constOrRegister(zero) == 0) {
             return jumpTo;
         }
         return DO_NOT_JUMP;
